Renamed `from` to `fromNode` and `to` to `toNode`
Replaced `diagram-drawer` with `diagram-drawer-advanced` and removed the latter.
