+++ src/demo/examples/complex.tsx
@@ -118,23 +118,23 @@ export const ComplexInstallation: FlowDiagram = {
         { id: "air", label: "Air", style: { strokeColor: "#2499E0", strokeStyle: "0" } },
     ],
     edges: [
-        { flow: "fuel", from: "input", to: "turbine-generator-set" },
-        { flow: "fuel", from: "input", to: "engine-generator-set" },
-        { flow: "emissions", from: "turbine-generator-set", to: "emissions-output" },
-        { flow: "emissions", from: "engine-generator-set", to: "emissions-output" },
-        { flow: "electricity", from: "turbine-generator-set", to: "pump-system" },
-        { flow: "electricity", from: "engine-generator-set", to: "pump-system" },
-        { flow: "electricity", from: "windfarm", to: "pump-system" },
-        { flow: "electricity-import", from: "power-from-shore", to: "pump-system" },
-        { flow: "electricity", from: "turbine-generator-set", to: "compressor-system" },
-        { flow: "electricity", from: "engine-generator-set", to: "compressor-system" },
-        { flow: "electricity", from: "windfarm", to: "compressor-system" },
-        { flow: "electricity-import", from: "power-from-shore", to: "compressor-system" },
-        { flow: "oil", from: "pump-system", to: "direct" },
-        { flow: "oil", from: "pump-system", to: "tabulated" },
-        { flow: "air", from: "compressor-system", to: "installation" },
-        { flow: "emissions", from: "installation", to: "emissions-output" },
-        { flow: "oil", from: "installation", to: "oil-output" },
-        { flow: "electricity-export", from: "installation", to: "electricity-output" },
+        { flow: "fuel", fromNode: "input", toNode: "turbine-generator-set" },
+        { flow: "fuel", fromNode: "input", toNode: "engine-generator-set" },
+        { flow: "emissions", fromNode: "turbine-generator-set", toNode: "emissions-output" },
+        { flow: "emissions", fromNode: "engine-generator-set", toNode: "emissions-output" },
+        { flow: "electricity", fromNode: "turbine-generator-set", toNode: "pump-system" },
+        { flow: "electricity", fromNode: "engine-generator-set", toNode: "pump-system" },
+        { flow: "electricity", fromNode: "windfarm", toNode: "pump-system" },
+        { flow: "electricity-import", fromNode: "power-from-shore", toNode: "pump-system" },
+        { flow: "electricity", fromNode: "turbine-generator-set", toNode: "compressor-system" },
+        { flow: "electricity", fromNode: "engine-generator-set", toNode: "compressor-system" },
+        { flow: "electricity", fromNode: "windfarm", toNode: "compressor-system" },
+        { flow: "electricity-import", fromNode: "power-from-shore", toNode: "compressor-system" },
+        { flow: "oil", fromNode: "pump-system", toNode: "direct" },
+        { flow: "oil", fromNode: "pump-system", toNode: "tabulated" },
+        { flow: "air", fromNode: "compressor-system", toNode: "installation" },
+        { flow: "emissions", fromNode: "installation", toNode: "emissions-output" },
+        { flow: "oil", fromNode: "installation", toNode: "oil-output" },
+        { flow: "electricity-export", fromNode: "installation", toNode: "electricity-output" },
     ],
 };
