+++ src/demo/examples/installation.tsx
@@ -88,14 +88,14 @@ export const installation: FlowDiagram = {
         { id: "oil", label: "Oil", style: { strokeColor: "#F75D36", strokeStyle: "0" } },
     ],
     edges: [
-        { flow: "electricity-import", from: "windfarm1", to: "installation1" },
-        { flow: "electricity-import", from: "windfarm2", to: "installation1" },
-        { flow: "electricity-import", from: "windfarm3", to: "installation1" },
-        { flow: "electricity-import", from: "power-from-shore", to: "installation1" },
-        { flow: "fuel", from: "input", to: "installation1" },
-        { flow: "emissions", from: "installation1", to: "emissions-output" },
-        { flow: "oil", from: "installation1", to: "oil-output" },
-        { flow: "electricity-export", from: "installation1", to: "electricity-output" },
+        { flow: "electricity-import", fromNode: "windfarm1", toNode: "installation1" },
+        { flow: "electricity-import", fromNode: "windfarm2", toNode: "installation1" },
+        { flow: "electricity-import", fromNode: "windfarm3", toNode: "installation1" },
+        { flow: "electricity-import", fromNode: "power-from-shore", toNode: "installation1" },
+        { flow: "fuel", fromNode: "input", toNode: "installation1" },
+        { flow: "emissions", fromNode: "installation1", toNode: "emissions-output" },
+        { flow: "oil", fromNode: "installation1", toNode: "oil-output" },
+        { flow: "electricity-export", fromNode: "installation1", toNode: "electricity-output" },
     ],
 };
 
@@ -163,14 +163,14 @@ export const installation2: FlowDiagram = {
         { id: "oil", label: "Oil", style: { strokeColor: "#F75D36", strokeStyle: "0" } },
     ],
     edges: [
-        { flow: "electricity-import", from: "windfarm1", to: "installation1" },
-        { flow: "electricity-import", from: "windfarm2", to: "installation1" },
-        { flow: "electricity-import", from: "windfarm1", to: "installation2" },
-        { flow: "electricity-import", from: "windfarm2", to: "installation2" },
-        { flow: "fuel", from: "input", to: "installation1" },
-        { flow: "fuel", from: "input", to: "installation2" },
-        { flow: "emissions", from: "installation1", to: "emissions-output" },
-        { flow: "oil", from: "installation1", to: "oil-output" },
-        { flow: "electricity-export", from: "installation1", to: "electricity-output" },
+        { flow: "electricity-import", fromNode: "windfarm1", toNode: "installation1" },
+        { flow: "electricity-import", fromNode: "windfarm2", toNode: "installation1" },
+        { flow: "electricity-import", fromNode: "windfarm1", toNode: "installation2" },
+        { flow: "electricity-import", fromNode: "windfarm2", toNode: "installation2" },
+        { flow: "fuel", fromNode: "input", toNode: "installation1" },
+        { flow: "fuel", fromNode: "input", toNode: "installation2" },
+        { flow: "emissions", fromNode: "installation1", toNode: "emissions-output" },
+        { flow: "oil", fromNode: "installation1", toNode: "oil-output" },
+        { flow: "electricity-export", fromNode: "installation1", toNode: "electricity-output" },
     ],
 };
