+++ src/demo/examples/other-example.tsx
@@ -42,7 +42,7 @@ const installation: FlowDiagram[] = [
         id: "installation-diagram-1",
         title: "Installation1",
         nodes: [{ id: "fuel-node" }, compressorSystem],
-        edges: [{ from: "fuel-node", to: "compressor-system1", flow: "fuel-flow" }],
+        edges: [{ fromNode: "fuel-node", toNode: "compressor-system1", flow: "fuel-flow" }],
         flows: [{ id: "fuel-flow", label: "Fuel", style: {} }],
         startDate: "2018-01-01",
         endDate: "2019-01-01",
@@ -52,8 +52,8 @@ const installation: FlowDiagram[] = [
         title: "Installation1",
         nodes: [{ id: "fuel-node" }, compressorSystem, compressor],
         edges: [
-            { from: "fuel-node", to: "compressor-system1", flow: "fuel-flow" },
-            { from: "fuel-node", to: "compressor1", flow: "fuel-flow" },
+            { fromNode: "fuel-node", toNode: "compressor-system1", flow: "fuel-flow" },
+            { fromNode: "fuel-node", toNode: "compressor1", flow: "fuel-flow" },
         ],
         flows: [{ id: "fuel-flow", label: "Fuel", style: {} }],
         startDate: "2019-01-01",
